+++ .gitignore
@@ -1,4 +1,4 @@
-.env
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
 /node_modules
