+++ package.json
@@ -1,18 +1,13 @@
 {
-  "name": "meal-planner-app",
+  "name": "react_table",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "bootstrap": "^5.1.3",
-    "history": "^5.1.0",
-    "npm-watch": "^0.11.0",
     "react": "^17.0.2",
-    "react-bootstrap": "^2.0.1",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.0.1",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.2"
   },
@@ -20,16 +15,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "watch": "npm-watch"
-  },
-  "watch": {
-    "build": {
-      "patterns": [
-        "src"
-      ],
-      "extensions": "js"
-    }
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
